[include/dxgmx/time.h] Change clock_t and time_t types to i32.

+++ include/dxgmx/time.h
@@ -7,8 +7,8 @@
 
 #define CLOCKS_PER_SEC 1000000
 
-typedef i64 clock_t;
-typedef i64 time_t;
+typedef i32 clock_t;
+typedef i32 time_t;
 
 struct tm {
    int tm_sec;
